retrying new firebase library
