+++ .github/workflows/android-build-publish.yml
@@ -1,4 +1,4 @@
-name: Android Apk - Github Release
+name: Android Apk - Build & Release
 
 on: [push]
 
@@ -8,7 +8,7 @@ jobs:
 
         strategy:
             matrix:
-                node-version: [12.x]
+                node-version: [10.x]
 
         steps:
             - uses: actions/checkout@v1
@@ -47,3 +47,17 @@ jobs:
                   automatic_release_tag: 'latest'
                   title: 'Beta Build'
                   prerelease: false
+
+            - name: Build Android Bundle to upload to Google Play
+              if: github.ref == 'refs/heads/master'
+              run: yarn bundle
+
+            - name: Publish Android Bundle to Google Play Store
+              uses: r0adkll/upload-google-play@v1
+              if: github.ref == 'refs/heads/master'
+              with:
+                  # The service account json private key file to authorize the upload request
+                  serviceAccountJsonPlainText: ${{ secrets.SERVICE_ACCOUNT_JSON_PLAIN_TEXT }}
+                  packageName: 'com.siristechnology.nepaltodayapp.beta'
+                  releaseFile: android/app/build/outputs/bundle/signedRelease/app-signed-release.aab
+                  track: internal
