--- .github/workflows/android-release-google-play.yml
@@ -1,44 +1,0 @@
-name: Android Apk - Google Play Release
-
-on: [push]
-
-jobs:
-    build:
-        runs-on: macos-latest
-
-        strategy:
-            matrix:
-                node-version: [12.x]
-
-        steps:
-            - uses: actions/checkout@v1
-
-            - name: Set up JDK 1.8
-              uses: actions/setup-java@v1
-              with:
-                  java-version: 1.8
-
-            - name: Use Node.js ${{ matrix.node-version }}
-              uses: actions/setup-node@v1
-              with:
-                  node-version: ${{ matrix.node-version }}
-
-            - name: Yarn Install, (Test)
-              run: |
-                  yarn
-
-            - name: Build Android Bundle to upload to Google Play
-              if: github.ref == 'refs/heads/master'
-              run: |
-                  sh rename_package.sh
-                  yarn bundle
-
-            - name: Publish Android Bundle to Google Play Store
-              uses: r0adkll/upload-google-play@v1
-              if: github.ref == 'refs/heads/master'
-              with:
-                  # The service account json private key file to authorize the upload request
-                  serviceAccountJsonPlainText: ${{ secrets.SERVICE_ACCOUNT_JSON_PLAIN_TEXT }}
-                  packageName: 'com.siristechnology.nepaltodayapp.beta'
-                  releaseFile: android/app/build/outputs/bundle/signedRelease/app-signed-release.aab
-                  track: internal
