+++ android/app/build.gradle
@@ -1,4 +1,6 @@
 apply plugin: "com.android.application"
+apply plugin: 'com.google.gms.google-services'
+apply plugin: 'com.google.firebase.crashlytics'
 
 import com.android.build.OutputFile
 
@@ -199,9 +201,7 @@ android {
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
     implementation "com.facebook.react:react-native:+"  // From node_modules
-    implementation "com.google.android.gms:play-services-base:16.1.0"
-    implementation "com.google.firebase:firebase-core:16.0.9"
-    implementation "com.google.firebase:firebase-messaging:18.0.0"
+    implementation "com.google.android.gms:play-services-base:16.1.0"    
     implementation 'me.leolin:ShortcutBadger:1.1.21@aar'
 
     if (enableHermes) {
@@ -211,9 +211,10 @@ dependencies {
     } else {
       implementation jscFlavor
     }
-
-
-
+    
+    implementation "com.google.firebase:firebase-messaging:20.1.7"
+    implementation 'com.google.firebase:firebase-analytics:17.4.1'
+    implementation 'com.google.firebase:firebase-crashlytics:17.0.0'
 }
 
 // Run this once to be able to run the application with BUCK
