+++ android/app/src/main/AndroidManifest.xml
@@ -16,13 +16,6 @@
     </activity>
     <activity android:name="com.facebook.react.devsupport.DevSettingsActivity" />
 
-    <service android:name="io.invertase.firebase.messaging.RNFirebaseMessagingService">
-      <intent-filter>
-        <action android:name="com.google.firebase.MESSAGING_EVENT" />
-      </intent-filter>
-    </service>
-    <service android:name="io.invertase.firebase.messaging.RNFirebaseBackgroundMessagingService" />
-
   </application>
 
 </manifest>
