+++ package.json
@@ -21,6 +21,9 @@
 		"@react-native-community/masked-view": "0.1.7",
 		"@react-native-community/netinfo": "5.6.2",
 		"@react-native-community/viewpager": "3.3.0",
+		"@react-native-firebase/app": "^7.0.1",
+		"@react-native-firebase/crashlytics": "^7.0.1",
+		"@react-native-firebase/messaging": "^7.0.1",
 		"apollo-cache-inmemory": "^1.6.6",
 		"apollo-client": "^2.6.9",
 		"apollo-link": "^1.2.14",
@@ -34,7 +37,6 @@
 		"node-fetch": "^2.6.0",
 		"react": "16.13.0",
 		"react-native": "0.61.5",
-		"react-native-firebase": "5.6.0",
 		"react-native-gesture-handler": "^1.6.0",
 		"react-native-localize": "^1.3.4",
 		"react-native-reanimated": "^1.8.0",
